refactor: cors middleware
